+++ src/goose/broker.clj
@@ -10,7 +10,7 @@
   (enqueue [this job] "Enqueues a Job for async execution.")
   (enqueue-batch [this batch] "Enqueues a Batch of Jobs for async execution.")
   (schedule [this schedule-epoch-ms job] "Schedules a Job for execution at given epoch-ms.")
-  (register-cron [this cron-opts job-description] "Registers a function for periodic execution in cron-jobs style.")
+  (register-cron [this cron-opts job-description] "Registers a function for recurring execution in cron-jobs style.")
 
   ;; worker/consumer
   (start-worker [this worker-opts] "Starts a worker process that does multiple things: