+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [0.5.3] - 16th-Sept-2024
 
 ### Added
-- View schedule and periodic jobs in console
+- View Scheduled and Cron jobs in console
 - Add dark theme
 - Poll jobs
 ### Changed